requirements.txt
asgiref==3.5.2
Django==4.0.6
django-cors-headers==3.13.0
django-environ==0.9.0
environ==1.0
python-dotenv==0.20.0
sqlparse==0.4.2
tzdata==2022.4

djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0